[ASR Refactor] Extract shared speech-to-text serving helpers
Move the mechanics of /v1/audio/transcriptions (form parsing/validation,
generate-request construction, dispatch, response serialization, adapter
resolution, duration probe, SSE lifecycle) into serve/speech_to_text.py;
transcriptions.py becomes a thin endpoint layer. Behavior-preserving:
existing transcription tests pass without touching any assertion.
build_transcription_generate_request gains task="transcribe" (default
unchanged). Prepares one shared seam for the in-flight auto-chunking
drafts and a sibling /v1/audio/translations endpoint.

Ref: #1267 (RFC scope: M5 follow-up refactor, no behavior change)
